make dashboard default team route
